requirements.txt
Django>=1.3
django-admin-tools>=0.4.0
django-debug-toolbar>=0.9.4
django-simple-captcha>=0.3.4
psutil
